AgenOps_UseCases/UseCase1/agentops_realtime_poc/requirements.txt
openai>=2.6.0
agentops>=0.4.21
python-dotenv>=1.0.0
